pyproject.toml
[project]
name = "tech-blog-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "google-genai>=1.68.0",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tech_blog_rag"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
markers = ["api: Gemini API を呼ぶテスト"]
